pyproject.toml
[project]
name = "aimemos"
version = "0.1.0"
description = "AI-based personal knowledge base service"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi>=0.121.1",
    "passlib>=1.7.4",
    "pocketflow>=0.0.3",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "uvicorn>=0.38.0",
    "python-multipart>=0.0.20",
    "python-docx>=1.1.0",
    "pypdf>=5.0.0",
]

[project.scripts]
aimemos = "aimemos.main:main"

[project.optional-dependencies]
demo = [
    "requests>=2.32.5",
]

[tool.uv]
package = true
